requirements.txt
apache-airflow-providers-apache-hdfs
apache-airflow-providers-docker
#apache-airflow-providers-apache-spark
pytest
wget
